casbin-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
casbin>=1.33.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
pydantic>=2.5.0
